pyproject.toml
[project]
name = "audiora"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "librosa==0.11.0",
    "matplotlib==3.10.8",
    "pyrtlsdr==0.4.0",
    "pyrtlsdrlib==0.0.4",
    "scipy==1.17.1",
    "sounddevice==0.5.5",
]
